txt2vid_inputs
